refactor: moved database in root directory
