+++ .gitignore
@@ -2,6 +2,7 @@
 build/
 coverage/
 node_modules/
+migrations/
 
 # Logs
 logs
