+++ package.json
@@ -37,13 +37,19 @@
     "check:lint": "npx eslint .",
     "check:markdown": "npx markdownlint \"**/*.md\"",
     "check:spell": "npx cspell lint --config cspell.json --no-progress --show-context \"**\"",
+    "db:format": "npx prisma format --schema schema.prisma",
+    "db:generate": "npx prisma generate --schema schema.prisma",
+    "db:generate:watch": "npx prisma generate --schema schema.prisma --watch",
+    "db:migrate": "npx prisma migrate dev --schema schema.prisma --name recluster",
+    "db:reset": "npx prisma migrate reset --schema schema.prisma",
+    "db:studio": "npx prisma studio --schema schema.prisma",
     "fix": "npx npm-run-all --npm-path npm --serial fix:*",
     "fix:database": "npm run format --workspace=packages/database",
     "fix:format": "npx prettier --write .",
     "fix:license": "npx license-check-and-add add -f license.json",
     "fix:lint": "npx eslint --fix .",
     "fix:markdown": "npx markdownlint --fix \"**/*.md\"",
-    "prepare": "npx husky install && npx ts-patch install -s && npm run generate --workspace=packages/database"
+    "prepare": "npx husky install && npx ts-patch install -s && npm run db:generate"
   },
   "devDependencies": {
     "@commitlint/cli": "^16.2.3",
@@ -71,6 +77,7 @@
     "markdownlint-cli": "^0.31.1",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.6.2",
+    "prisma": "^3.12.0",
     "ts-node-dev": "^1.1.8",
     "ts-patch": "^2.0.1",
     "tsconfig-paths": "^3.14.1",
