--- packages/database/.gitignore
@@ -1,2 +1,0 @@
-# Prisma
-migrations/
