--- packages/database/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@recluster/database",
-  "version": "0.0.1",
-  "description": "reCluster database",
-  "author": "Carlo Corradini <carlo.corradini98@gmail.com>",
-  "license": "MIT",
-  "private": true,
-  "main": "build/index.js",
-  "types": "build/index.d.ts",
-  "scripts": {
-    "format": "npx prisma format --schema schema.prisma",
-    "generate": "npx prisma generate --schema schema.prisma",
-    "generate:watch": "npx prisma generate --schema schema.prisma --watch",
-    "migrate": "npx prisma migrate dev --schema schema.prisma --name recluster",
-    "reset": "npx prisma migrate reset --schema schema.prisma",
-    "studio": "npx prisma studio --schema schema.prisma"
-  },
-  "peerDependencies": {
-    "@prisma/client": "3.12.0"
-  },
-  "devDependencies": {
-    "@prisma/client": "3.12.0",
-    "prisma": "^3.12.0"
-  }
-}
