+++ subgraphs/nodes/package.json
@@ -12,7 +12,6 @@
   "dependencies": {
     "@prisma/client": "^3.12.0",
     "@recluster/configs": "^0.0.1",
-    "@recluster/database": "^0.0.1",
     "@recluster/errors": "^0.0.1",
     "@recluster/graphql": "^0.0.1",
     "@recluster/helpers": "^0.0.1",
