+++ tsconfig.json
@@ -2,7 +2,6 @@
   "files": [],
   "references": [
     { "path": "packages/configs" },
-    { "path": "packages/database" },
     { "path": "packages/errors" },
     { "path": "packages/graphql" },
     { "path": "packages/helpers" },
